AI/requirements.txt
flask==2.3.3
tensorflow==2.19.0
keras==2.13.1
numpy==1.24.3
pillow==9.5.0
protobuf==4.21.12 